Refactor getChangeCosts
The method is changed such that it has two output parameters. There is the
rule to avoid output parameters. But in this case adding the output
parameters is I think acceptable as it drastically reduces the complexity
and the number of HashMaps created during execution. Also the method is
private, which means the chance for the method to be misunderstood or used
wrongly is quite low.